+++ src/base/cs_solve_all.cpp
@@ -42,6 +42,7 @@
 
 #include "atmo/cs_atmo_1d_rad.h"
 #include "atmo/cs_atmo.h"
+#include "atmo/cs_atmo_soil_model.h"
 #include "base/cs_1d_wall_thermal.h"
 #include "base/cs_ale.h"
 #include "base/cs_array.h"