+++ NEWS.md
@@ -57,6 +57,11 @@ Master (not on release branches yet)
 
 ### Physical modelling
 
+- Add a new plant model in the soil model. See J. Vernier et al. 2026
+  "A Soil-Plant-Atmosphere Continuum model coupled to CFD to simulate plant
+  energy and water exchanges in heterogeneous microclimates" for further
+  information.
+
 - Refurbish 1D atmospheric radiative model.
 
 - Allow the usage of Cooling tower module with Lagrangian particle tracking